Random Room Generation + Navigating rooms
Added DungeonLayout Class which generates a certain amount of rooms in the dungeon and stores them in a map together with an integer-key which represents the room's position in the dungeon.

Rooms are categorized by an enum and have fixed names and background images. The first Room in the Dungeon is always the "Entrance" and this type of Room will not be randomly generated elsewhere.

The GameScene Class has an explorationView() which displays buttons that can be used to travel to adjacent existing rooms and will throw an error message, if there is no room in a chosen direction.