Inventory Update, Restructured Entity Class and derived Classes
Equip and Drop Button in Inventory are mostly functional.
Equipped Weapon is highlighted in Inventory.
When Weapon gets equipped, player stats (damage and accuracy) are updated accordingly.
Accuracy now actually matters in the attack method.
Added starting Weapon (Dagger) to the Inventory.
At the moment, a new Weapon is found when killing an enemy (no visual info for the player yet)